fixed api links

+++ WebApi/Models/Game.cs
@@ -217,7 +217,7 @@ namespace WebApi.Models
                 Random r = new Random();
                 price = r.NextDouble() * player.Money;
                 if (player.SpendMoney(price))
-                    this.Units.Add(this.CharacterServices.GetWeirdo(price, player.Side));
+                    this.Units.Add(this.CharacterServices.GetWeirdo((int)price, player.Side));
             }
 
             price = this.CharacterServices.GetPrice(name, player.Team);
